Allow WAL implementation to be passed as an option
Allow users of the TSDB library to pass an implementation of WAL.
Defaults to use SegmentWAL if the option is nil.

For example, this allows the WAL to be disabled, which can be desirable
in situations where data integrity is safeguarded by other mechanisms
(e.g. by reading from a message bus), or when the data being written is
not valuable enough to warrant the I/O overhead and storage used by the
WAL.

No changes are required for existing users.